+++ src/Actions/DeleteDomainCertificate.php
@@ -17,13 +17,18 @@ readonly class DeleteDomainCertificate
      * @throws \Illuminate\Http\Client\ConnectionException
      * @throws \Illuminate\Http\Client\RequestException
      */
-    public function handle(int $serverId, int $siteId, int $domainRecordId): bool
-    {
+    public function handle(
+        int $serverId,
+        int $siteId,
+        int $domainRecordId,
+        int $certificateId
+    ): bool {
         $path = $this->client->path(
-            '/servers/%s/sites/%s/domains/%s/certificate',
+            '/servers/%s/sites/%s/domains/%s/certificates/%s',
             $serverId,
             $siteId,
-            $domainRecordId
+            $domainRecordId,
+            $certificateId
         );
 
         $response = $this->client->delete($path)->throw();