+++ src/Data/Certificate.php
@@ -3,6 +3,7 @@
 namespace SebastianSulinski\LaravelForgeSdk\Data;
 
 use Carbon\Carbon;
+use SebastianSulinski\LaravelForgeSdk\Data\Concerns\HasLinks;
 use SebastianSulinski\LaravelForgeSdk\Enums\Certificate\KeyType;
 use SebastianSulinski\LaravelForgeSdk\Enums\Certificate\RequestStatus;
 use SebastianSulinski\LaravelForgeSdk\Enums\Certificate\Status;
@@ -11,8 +12,12 @@ use SebastianSulinski\LaravelForgeSdk\Enums\Certificate\VerificationMethod;
 
 readonly class Certificate
 {
+    use HasLinks;
+
     /**
      * Certificate constructor.
+     *
+     * @param  array<string, mixed>  $links
      */
     public function __construct(
         public int $id,
@@ -22,6 +27,8 @@ readonly class Certificate
         public ?VerificationMethod $verificationMethod = null,
         public ?KeyType $keyType = null,
         public ?string $preferredChain = null,
+        public ?bool $active = null,
+        public array $links = [],
         public ?Carbon $createdAt = null,
         public ?Carbon $updatedAt = null,
     ) {}